pyproject.toml
[tool.poetry]
name = "wrtn-assignment"
version = "0.1.0"
description = ""
authors = ["youngsupchoi <sup1214@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
openai = "^1.54.4"
python-dotenv = "^1.0.1"
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
datasets = "^3.1.0"
pinecone-client = "^5.0.1"
pinecone-text = "^0.9.0"
pinecone-notebooks = "^0.1.1"
langchain = "^0.3.7"
langchain-community = "^0.3.7"
langchain-text-splitters = "^0.3.2"
langchain-openai = "^0.2.9"
nltk = "^3.9.1"
tiktoken = "^0.8.0"
langchain-pinecone = "^0.2.0"
kiwipiepy = "^0.20.1"
rank-bm25 = "^0.2.2"
tqdm = "^4.67.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
